Popup overflow
Moving `#hotkey-info` inside of `#installed` still makes the most sense to me. With a few adjustments to the CSS, we can lose the js calc for its height altogether. I also don't see a reason for it to be the entire height of the popup if the popup is already tall enough to accommodate it, so I adjusted that as well. I don't think a wrapper would be a huge deal, but I don't see a need for one anyway.

I suppose there may be scenarios I'm not accounting for, but everything appears to look right and work correctly with these changes. Goes without saying, whenever I modify js it should be scrutinized. In this case I wouldn't be surprised if I left some code made irrelevant by what I removed.